if the domain doesn't have cert skip

+++ singbox/run.sh
@@ -71,10 +71,13 @@ for TUIC_DOMAIN in ${TUIC_DOMAINS//;/ }; do
 	IFS=':' read -ra PARTS <<<"$TUIC_DOMAIN"
 	domain=${PARTS[0]}
 	port=${PARTS[1]}
-	cp configs/05_inbounds_4010_tuic.json configs/05_inbounds_4010_tuic_$port.json
-	sed -i "s|tuic_in|tuic_in_$port|g" configs/05_inbounds_4010_tuic_$port.json
-	sed -i "s|4010|$port|g" configs/05_inbounds_4010_tuic_$port.json
-	sed -i "s|SERVER_NAME|$domain|g" configs/05_inbounds_4010_tuic_$port.json
+	crt=/opt/hiddify-server/ssl/$domain.crt
+	if [[ ! -f $crt ]] && [[ ! -f ${crt}.key]]; then
+		cp configs/05_inbounds_4010_tuic.json configs/05_inbounds_4010_tuic_$port.json
+		sed -i "s|tuic_in|tuic_in_$port|g" configs/05_inbounds_4010_tuic_$port.json
+		sed -i "s|4010|$port|g" configs/05_inbounds_4010_tuic_$port.json
+		sed -i "s|SERVER_NAME|$domain|g" configs/05_inbounds_4010_tuic_$port.json
+	fi
 done
 rm configs/05_inbounds_4010_tuic.json
 
@@ -84,10 +87,12 @@ for HYSTRIA_DOMAIN in ${HYSTRIA_DOMAINS//;/ }; do
 	IFS=':' read -ra PARTS <<<"$HYSTRIA_DOMAIN"
 	domain=${PARTS[0]}
 	port=${PARTS[1]}
-	cp configs/05_inbounds_4100_hysteria.json configs/05_inbounds_4100_hysteria_$port.json
-	sed -i "s|hysteria_in|hysteria_in_$port|g" configs/05_inbounds_4100_hysteria_$port.json
-	sed -i "s|4100|$port|g" configs/05_inbounds_4100_hysteria_$port.json
-	sed -i "s|SERVER_NAME|$domain|g" configs/05_inbounds_4100_hysteria_$port.json
+	if [[ ! -f $crt ]] && [[ ! -f ${crt}.key]]; then
+		cp configs/05_inbounds_4100_hysteria.json configs/05_inbounds_4100_hysteria_$port.json
+		sed -i "s|hysteria_in|hysteria_in_$port|g" configs/05_inbounds_4100_hysteria_$port.json
+		sed -i "s|4100|$port|g" configs/05_inbounds_4100_hysteria_$port.json
+		sed -i "s|SERVER_NAME|$domain|g" configs/05_inbounds_4100_hysteria_$port.json
+	fi
 
 done
 rm configs/05_inbounds_4100_hysteria.json
